# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dictionary App for IT",
        "description": "API Docs for Dictionary",
        "contact": {
            "email": "quanghuybest@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/get-suggest-all": {
            "get": {
                "tags": [
                    "Get Suggest"
                ],
                "summary": "Lấy tất cả từ gợi ý khi tìm kiếm",
                "operationId": "8ab4efea9faa557af84fa898027c4d02",
                "responses": {
                    "200": {
                        "description": "Lấy thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy từ tất cả từ vưng thành công !"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Firewall"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy từ gợi ý",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Hiện tại chưa có gợi ý!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/get-suggest": {
            "get": {
                "tags": [
                    "Get Suggest"
                ],
                "summary": "Lấy danh sách các từ gợi ý theo chuyên ngành cụ thể",
                "operationId": "b7a78547be9280248854260c7e079b9d",
                "parameters": [
                    {
                        "name": "specialization_id",
                        "in": "query",
                        "description": "Nhập id của chuyên ngành",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách các từ gợi ý theo chuyên ngành",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "suggest_name": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Ví dụ: 1"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy từ gợi ý",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Lỗi rồi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}